loginApp/requirements.txt
asgiref==3.6.0
Django==4.1.6
psycopg-binary==3.1.8
psycopg2-binary==2.9.5
PyJWT==2.6.0
sqlparse==0.4.3
typing_extensions==4.4.0
